Fix: unable to instantiate translator using alias
The TranslationServiceProvider from Laravel is a deferred provider, which does not register until the 'translator' or 'translation.loader' bindings are needed. Laravel does not take aliases to these string-bound bindings into consideration when deferring the provider. Since it's a bit hard to follow all of these aliases we now use Laravel's string-bound bindings in this package as well.