chore: solve PHP explicit nullable type deprecation
Previously attempted in 7db1d2ce3478eebfb15bee2f7aca26c7e44a0a08, but that did not include a fix for all places where this was occurring. Although we're changing the method signature of an interface, this won't be a BC. This is the case because the parameter was implicitly nullable before. See https://github.com/laravel/framework/issues/50803